value meta size 8 bytes -> 2 bytes
git-svn-id: svn+ssh://code.etolabo.org/usr/etolabo/var/svn/ficia/kumofs/trunk@1163 b300df90-00d7-44b1-9271-3fa3ef26909a

+++ src/logic/gateway/scope_store.cc
@@ -120,7 +120,7 @@ void scope_store::Set(void (*callback)(void*, set_response&), void* user,
 		const char* key, uint32_t keylen, uint64_t hash,
 		const char* val, uint32_t vallen)
 try {
-	uint64_t meta = 0;
+	uint16_t meta = 0;
 	if(!life) { life.reset(new msgpack::zone()); }
 	rpc::retry<server::proto_store::Set>* retry =
 		life->allocate< rpc::retry<server::proto_store::Set> >(
